+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+Features
+
+- #23: Allow user to configure cancel_previous_builds value for each workflow which is set to false per default
+
 ## Version 3.2.0
 
 Features